Rename validate-json target as pep8 and improve it
When discussing with the infra guys they have mentioned it would be
easier to call our linting job pep8, it's indeed badly named but that
target has been used all over openstack for linting projects. As a bonus
point it would make things easier to add the job to the gate. To make
that patch much more interesting than a three characters change I have
improved the validate-samples script to detect if jsonlint was present
and if not fallback to the standard python -mjson.tool which give you
less details but nonetheless works if jsonlint is present.

Change-Id: I8d71a229917004dfd7223a16e4f270101cf2f0a8