add better validation checks (part 1)
This patch adds a script for validating YAML files, and replaces the
existing JSON checks with one largely identical to the YAML check.

This also provides a script that can be installed as a pre-commit hook
that will perform this checks when you commit changes.  You can install
the hook by running tools/pre-commit-hook --install.

Change-Id: Ib4742a9db062362cfa61d669c691151bc1ca376c