Use hostPort to lock 1 nova-compute/nova-network per node
Kubernetes does not have a horizontal scaling feature that works per
node.  A workaround is to use hostPort to cause the scheduler not to schedule
several replicationController replicas to each node.  By doing so, only one
nova-compute is launched per node up to a total of 6 nodes which is the most
we can generate in testing.

Change-Id: I1621f15cf1ec502418b454300645b46bc796dc8b