Fixed Neutron Service Proxy and Simplified Naming
Previously, the k8s neutron-server service was using a selector
for the neutron pod that did not exist. The selector name defined
in the service yaml was more appropriate than the label (neutron)
defined in the pod yaml, so the label was updated in the pod
yaml.

Additionally, the naming convention for the neutron pod was
changed from neutron-controller to neutron-server to more
correctly depict the pods functionality.

Change-Id: I99465069bb3c37e7bf17c6d7a8f3559f87bf7140