Containerize OpenStack based upon SPC and fig
This specification proposes using fig (being renamed to compose
soon) to provide single-node multi-container orchestration.  By
using this mechanism, a very simple Ansible Playbook could easily
deploy a single node in to a specific role type - such as a controller
node, a compute node, or a storage node.

This specification further proposes using super-privileged containers
to provide solutions for the upgrade and rollback use cases of an
OpenStack deployment.

Change-Id: I56ff1fdf8b19b47be97778b55ea947ebb43995c1