Complete docker compose setup.
This change proposes a minimal openstack deployment stood up using
docker-compose.  Included is a script to generate the necessary
environment variables for use by the containers as well as an
'openrc'.  There is also a script which demonstrates starting the
included services.

We have this working locally using patched containers and can start
instances with networking etc.  However, there are several changes
required to various containers that will need to be made to get this
working.  We are also using Steven Dake's patches to docker-compose to
allow the use of pid=host.  I propose we either merge this change or
build changes on top of this to get a functioning installation.

Co-authored-by: Ryan Hallisey <rhallise@redhat.com>

Change-Id: I5e6051054d31edf9840c3a22502d4a7963893f29