Add a "admin" user via crux to keystone container
The keystone container was changed to only add a keystone user.  The rest
of the infrastructure in Kolla expects the name "admin" from a "can I
use a client to access the APIs perspective."  As a result, we need a user
for keystone and a user for admin which this patch adds.

Change-Id: I3a3e938f946df1fd5982e89397ef61368f1f3674