Tell mariadb to restart
If a system is rebooted with kolla running, all the services will
continually restart.  There appears to be some locking problem in
docker which will result in docker imploding.  One problem is a
key dependency, mariadb, is never started on a reboot because it is
missing restart: always.  Also restart the mariadb data container
until it returns success.

Change-Id: I599d28548ff041712d54a702c80fe36c321adf7f