libvirtd provides was removed from centos7.1
the libvirtd package "Provides" wwas removed from the libvirt packaging.
To workaround this problem just install libvirt-daemon in CentOS 7.1 or Fedora
20+.

WIthout this change, an install of libvirtd does not actually install
libvirtd.  The yum tool will not fail if libvirtd is specified so we weren't
catching this in the build.

Change-Id: I81e64f6f659aa0466369dad4b0574ebb24c99555