Remove building of pam+noaudit and place in copr instead
COPR is a repository for third party packages built by the Fedora build
system.  Instead of dragging in 250MB of dependencies to build pam, build
externally and load only the RPM into the image.

The pam that is built externally is built with the --with-noaudit flag
and a Provides: pam line.

Change-Id: I28396996943d2fdc12b253faf62af8623c8b0e73