Do not bindmount nova-compute directories
Nova uses a data container.  Data containers are expected to
use the VOLUME keyword and then use volumes_from to access
those volumes.  Throwing in a bindmount to the host filesystem
results in unpredicble behavior, usually in the form of files
being stored on the host filesystem that should be persisted
in the compute data volume.

Change-Id: I60a7873995c9397368ed4df61d5a6d4e8250f3c2