The neutron dockerfile needs an update for Kilo
policy.json comes from the openstack-neutron package and is required
for the agents to run.  So install it.  fwaas_driver.ini has been deleted
in master and in the kilo branch, but the agents won't run without the
file present.  So copy one from the Juno release.

fwaas_driver.ini is probably not actually required, but there is a
configuration problem somewhere in our Neutron code base.  We are out
of time for the Kilo branch, so this short term workaround should be fine.

Change-Id: I154599925308765349634b1a53e598599fc20ec1
Partially-Implements: blueprint port-kilo