Fixes Nova NOVNC and Consoleauth Startup Failure
Previously these services would not start because the start
script requires the associated log file to exist. The default
settings for these 2 log files are blank causing the failure.
Since the log files are not required for the services to start,
this patch removes the parameter requirement from both start
scripts.

Change-Id: I84d3d128728075e52282b49d9df229917a34a7b5