Allow private docker registries
Change the sed seperator to accept a namespace with a private registry
and add an option to specify the registry. Below is an example for a
registry located at 192.168.0.10:5000.

$ ./build-all-docker-images -r 192.168.0.10:5000

The final image name would be something like this:
'192.168.0.10:5000/kollaglue/centos-rdo-base'

All build, push, and pull commands will still function the same.

Change-Id: Ic0ff96bb8119ddfab15b99e9f1e21cfe8d321dab