Fixes Nova NoVNC
Previously, vnc was not completely working properly because
nova.conf config flags were not properly implemented in t he correct
containers.

This patch moves the nova.conf config flags for vnc to the proper
containers. Additionally, the patch modifies the proxy client ip
param to be consistent with the param within nova.conf and adds
all the novnc params to docs.

Change-Id: I1944baca64ccd113bbca582784b08f90dc0eff69