Always provide a tag when building with build-all-docker-images
When we're not providing a tag, the build script will use one based on
the short SHA of the tip of the current git branch.

Make the `build-all-docker-images` script always provide a tag to
`build-docker-image` in order to prevent issues with Docker not being
able to find the base image when changing branches during an ongoing
build.

Closes-Bug: #1458739
Change-Id: Id6925b792f46d14b20a7dbebfaf0d351ddec6538