Auto generate minimal environment variables doc
This script will generate the minimal-environment-variables doc
based on the what each services has in the check_required_var field.
Anytime that field is changed run this script and commit the changes.
Since Neutron is a thick container, the script will get a few duplicates
in the neutron-agents field.  When Neutron is thinned out, this script
will function normally.

Closes-bug: #1459347

Co-Authored-By: Lon Hohberger <lhh@redhat.com>
Change-Id: I1ba6bd5148af12cacfd872591b3c7b84a8dccba2