Make crux compatible with delorean packaging
The previous version of crus 6-1 installed pbr hooks.  This is
incompatible with CentOS RDO.  Remove the requirements.txt file
from the RPM that is installed.  This required placing the files
in a new COPRS repository.  Since Lars no longer contributes to
Kolla, I imported his src rpm, made the necessary adjustments,
and created a new COPRS repo for it.

The result of the PBR hooks is that any container that ran crux
would fail with a distribution error about argparse not being
present when it was present.

Change-Id: Ibdbccd9e0dd5aeb6d6292bd633b9d2dde8c7b4ce
Closes-bug: #1460886