Ansible multi-node specification
This specification outlines some of the design goals and practices to
be used to implement an Ansible multi-node deployment tool.  The goals
of this work are simplicity, optimizing deployment time, and Operator
flexibility.

The long term objective is for Kolla to become the de-facto solution
for deploying OpenStack in containers.

Partially Implements: blueprint ansible-multi
Co-Authored-By: Sam Yaple <sam@yaple.net>

Change-Id: I8c5d8b7e69c4ed54624061428e2fbde6af6a1815