Fix Heat container env vars and dependencies
This commit adds a few dependencies (that are actually packaging errors)
present in all of the containers, as well as specifically in
heat-engine. Some of the environment variables were not set correctly in
the heat-api-cfn container. The environment variables have also been
correctly added to the genenv script.

Also removed duplicated compose file data already present in
heat-api-engine.

Closes-Bug: #1461872
Change-Id: I9631d1bc6ea42425272570a9dda8cbf5edcef4f3