Add openvswitch container
Add openvswitch containers. Openvswitch is seperated into two running
containers:
  ovs-db-server
  ovs-vswitchd

The db container is what everything interacts with. Neutron tells the db
its information, the vswitchd container reads this db and talks to the
kernel module.

The db does not need to be persistent since the plugin agent will
repopulate all the information each time it is restarted and the plugin
agent is required to be restarted when the ovs-vswitchd container is
stopped/started to ensure the ports are setup properly.

This container requires /run for the socket and /lib/modules:ro to load
the appropriate host kernel module.

This userspace tools and the kernel module do _not_ have to match
versions. Additionally, even though it is recommended that the userspace
tool be newer than the kernel version to take advantage of all the
features, it is not required.

Partially Implements blueprint: openvswitch-container

Co-Authored-By: Sam Yaple <sam@yaple.net>

Change-Id: I70e3807be32c9a07346d316e7856421ecf468b9a