Add base image for oraclelinux
This patch provides the basis for an alternate set of images to be
available in Kolla, based on oraclelinux.

It is difficult ahead of time to know all base dependencies and exactly
which repos they'll come from, so at this time I have just added the bare
minimum.

In the next commit keystone will require some essentials that will
likely make sense to add here.

NOTE: To build this image you must set the following .buildconf in the
root of the kolla repo:

cat >> .buildconf << EOF
PREFIX=ol-binary-
EOF

Partially implements: blueprint add-optional-oraclelinux-images

Change-Id: I1ccce8eebe8fa0d4342af2aa482c083d074bda66