Have test retry building failed images
The image build test now retries 3 times before declaring an image as
failing. This should help the gate in case of unreliable network
connectivity to the mirrors.

We expect the penalty of retrying to build failed images to be low
thanks to docker cache. Only the failing layer should be retried.

Change-Id: I15fd4926019590a150eb22dc885d0ef8cd31e7cc
Closes-Bug: #1466677