Update centos-source-base from symlink to a real layer
There are additional depedencies on top of the binary base that are
needed to build and install source components.

This commit moves the base image for source installs from a symlink to
it's own layer, and adds the dependencies that will be needed to install
keystone (more can be added as we add new components).

Partially-Implements: blueprint install-from-source

Change-Id: I56ef6eafb95530e975988dffb62606309dba274e