Remove RDO dependency from source images
It turns out the only dependencies we need from RDO in the source builds
are python-pip, python-cliff for crux, and crudini, all of which can be
pulled from pypi.

We should remove the dependency on RDO for these builds which is one of
the reasons for source builds in the first place.

Change-Id: I11e26e96e6c94a5df5db48204714229bbd85a9aa
Closes-Bug: #1476760