Add new test script for parallel builds
This adds a new test script for the new python based build script. A new
tox target has been created to execute it for usage in the gate.
Eventually, the other build script stuff can be deleted, but that won't
be done until full functionality exists in the new build.py.

Change-Id: Id9dd19926b1b9d875dcb069fd13d6450318374c8
Paritally-Implements: blueprint build-script