Prepare for MD to RST conversion
Sphinx and OpenStack documentation standards require the use of
RST formatted documentation.  In order to get out docs integrated
into docs.openstack.org, the following steps must be taken:

1. we need to first convert the documentation to RST
2. sphinx support must be added to the repository
3. project-config must be updated with a publish-docs job

The ability to pubish to docs.openstack.org is a privilege only
available to Big Tent projects.  As a result, we should be taking
advantage of this to help spread the word on our implementation.

Note the documentation was converted with a tool called pandoc
the syntax of which was:
pandoc -f markdown -t rst file.md > z
mv z file.rst

In order for git to preserve history the files have to be git mv'ed
in one commit followed by the pandoc operation in a separate commit.

Change-Id: Iba2a70b989e8305da03e8204a9b130d457b00cf0