Rework documentaton to actually get people rolling
Individuals are beginning to evaluate Kolla and our quickstart
documentation is in terrible terrible shape.  Pleae just approve
this change but do leave feedback.  I will address changes in
an iterative process in future patches rather then having a
perfect documentation patch in one bang.

Since this isn't code, there is not a high cost to refactoring
or risk of regressions.

Change-Id: I222cf15f4c5d748fb2daf3638094b4e7b35f8844