Replace config-external with a JSON file
Replace config-external with JSON file. The JSON file will
be placed in each of the services directories with expected
location and destination.

Set-configs.py will be responsible for interpreting the JSON file,
creating the necessary directories, moving config files, and
and providing a command line for start.sh to run the service
with the correct config files specified.

Partially-Implements: blueprint replace-config-external
Change-Id: I5e2e69dfe3ae7f938fcf51f1cd450aaa10e7f1e3