Make logging more better
build.py logging has never done what people think it should do, this
patch aims to fix that. All logs related to an image will now be
printed with the image name prefixed to the message.

To acheive this we have to print out the messages as we get them, even
in parallel building mode. If you wish to view the logs for an single
image, a grep would be advised to filter the logs.

Change-Id: I15ae361892f06a47afafe73750ddd8628697cfea
Implements: blueprint build-script