Start of removal of docker directory
Start off the removal of the docker directory by making the gate
only run against docker templates.  The idea is we are going to
from this patch set forward completely abandon the functioning of
the docker directory and focus on making templates work properly.

In order to facilitate that, this patch set makes sure each change
gates properly for the changed service in the Templates build.

Note because git review and gerrit can't keep history on removal
followed by a git move operation, we first git rm the files affected
then git mv them to get the gate working again.

Every other patch in this patch set will fail the gate.  That is
expected behavior.

Change-Id: I1be2e2638aef4ada038bfe5f3dd563f040542df2