Install which if its not installed in base
Some variants of RHEL containers are missing the 'which' command.
Percona's software used by mariadb depends on which, but doesn't have
a hard dependency requirement since it is usually installed in the
distribution.  To make mariadb work multinode on RHEL, install which
all the time for RPM distros.

Change-Id: I2c6aaa77a1fc8b7028b658a0dfd1935c1f1d738e
Partially-Implements: blueprint rhel-based-image-support