Remove debug output in base image yum command
The base image contains -d 10 in a yum command which prints unhelpful
debug output for one of the yum commands.  Remove the extranous debug
output.

backport: liberty

TrivialFix

This patch is a trivial fix and is a test to set boundaries on what
a trivial fix looks like that doesn't require a bug tracker entry.  If
reviewers feel a trivial fix isn't documented properly by this commit
message, please weigh in, as this is a slight suggested change to our
policy as first institued by the Magnum community and brought to my
attention by Jeff Peeler.  Hopefully this compromise won't be abused :)

Change-Id: I57f00978d0baaefd73592f4ad6a655e663d83478