Install agent dependencies from source for CentOS
Several packages are not installed from RPM packages which are
required in order for CentOS from source to operate correctly.

This looked like sudo was failing to execute, but it was because
the dnsmasq binary was not found.

backport: liberty

Change-Id: If23cfbacd55d9089ffec442399ee0151adea923f
Closes-Bug: #1503412
(cherry picked from commit bc0f8b422ad8a49d59ab2a504a108554cc75cbc0)