Release version 1.0.0
Our current version based upon conversations with TTX was
0.1.0. It appears now that the full API is nearly stable.  It is
stable enough to call our first Liberty release 1.0.0.

The API consists of the following data:
The locations of config files passed to the container
The environment variables expected by the container
The JSON encoded file containing the copy and cmd operations

Each container in current master follows the same sytnax and
style for all of the implementations of every container that
atleast has been Ansibilized.

backport: liberty

Change-Id: I2064689ca6fd9f332ad26624e1b93f6aaa08e79b
Closes-Bug: #1503494