Remove compose from the repostiory since its dead code
Several people over the liberty cycle committed to maintaining
compose, but it was never completed.  Rather than carry compose
in our repoistory until the end of time in a broken state I propose
removing it.  If someone comes along and wants to resseruct the code
and make it functional in the future, I'd be 100% open to merge a
patch in the future which reintroduces this functionality.  In the
meantime it is confusing to users which ask me on a daily basis if
compose is supported which I have to go into a 10 minute monologue
about why its in the repository.

Fortunately git contains full history so we can always revert this
change at a later time if someone wants to actually fixt his code
base up and make it useable in some way.  I think the proper answer
is a seperate repoistory (kolla-compose) which contains any compose
related bits.

Please consider removing since nobody is using this code and its
dead.

TrivialFix

backport: liberty
Change-Id: I8b7c39b00d1967184d363af7db86d1ff8b5e725f