Increase Ceph RPM gpg key timeout to 90 seconds
The default timeout for Ceph GPG rpm key retrieval is 30 seconds.
In my testing, the GPG key takes approximately 50 seconds to download
often resulting in a failure to build containers that need to retrieve
the Ceph GPG sign key.  Crank up the timer to 90 seconds so the key
is more likely to be downloaded, allowing images to be built.

backport: liberty

TrivialFix

Change-Id: I7420cdf8d3b61aa9f4f52795fccbe5da3e48d57b