Make Ceph install from Source on RPM
Ceph doesnt install from source on RPM because Ceph pulls in
python-requests.  This causes a cpio error during unpacking
because requests is installed in the base image.  A simple
workaround is to remove in the nova-compute container which
is the only container this conflict occurs within.

backport: liberty

Change-Id: I493900854a53c98f1a5cb6478675a18feceefea5
Closes-Bug: #1508236
(cherry picked from commit 194bb93ff26cfda92ff79742d361350a353ecacd)